les 4 opdracht 3

+++ les4/pe4_3.py
@@ -1,5 +1,5 @@
 uurloon =  input('wat verdien je per uur: ')
 uren = input('Hoeveel uur werk je: ')
-salaris =  int(uurloon) * int(uren)
+salaris =  float(uurloon) * float(uren)
 line1 = '' + str(uren) + ' uur werken levert je ' + str(salaris) + ' euro op '
 print(line1)
